requirements.txt
black==24.10.0
click==8.1.7
colorama==0.4.6
contourpy==1.3.0
cycler==0.12.1
distlib==0.3.9
et_xmlfile==2.0.0
filelock==3.16.1
flake8==7.1.1
fonttools==4.54.1
iniconfig==2.0.0
kiwisolver==1.4.7
matplotlib==3.9.2
mccabe==0.7.0
mypy-extensions==1.0.0
numpy==2.1.3
openpyxl==3.1.5
packaging==24.1
pandas==2.2.3
pathspec==0.12.1
pillow==11.0.0
platformdirs==4.3.6
pluggy==1.5.0
pycodestyle==2.12.1
pyflakes==3.2.0
pyparsing==3.2.0
pytest==8.3.4
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
six==1.16.0
tzdata==2024.2
virtualenv==20.27.1
